fix: make sync_http.cjs self-contained so wasm-bindgen-test and bundlers get the worker
